@@ -3,7 +3,7 @@
   "version": "0.0.0",
   "scripts": {
     "ng": "ng",
-    "start": "ng serve",
+    "start": "ng serve demo",
     "start_manager": "ng serve manager --port 4200",
     "start_doc": "ng serve demo --port 4201",
     "start_admin": "ng serve admin --port 4202",
@@ -18,17 +18,17 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^13.3.1",
+    "@angular/animations": "^14.3.0",
     "@angular/cdk": "^13.3.2",
-    "@angular/common": "^13.3.1",
-    "@angular/compiler": "^13.3.1",
-    "@angular/core": "^13.3.1",
-    "@angular/forms": "^13.3.1",
+    "@angular/common": "^14.3.0",
+    "@angular/compiler": "^14.3.0",
+    "@angular/core": "^14.3.0",
+    "@angular/forms": "^14.3.0",
     "@angular/material": "^13.3.2",
     "@angular/material-moment-adapter": "^13.3.1",
-    "@angular/platform-browser": "^13.3.1",
-    "@angular/platform-browser-dynamic": "^13.3.1",
-    "@angular/router": "^13.3.1",
+    "@angular/platform-browser": "^14.3.0",
+    "@angular/platform-browser-dynamic": "^14.3.0",
+    "@angular/router": "^14.3.0",
     "@codeffekt/ce-barcode": "^13.9.0",
     "@codeffekt/ce-code-editor": "^13.9.0",
     "@codeffekt/ce-core-data": "^1.1.2",
@@ -54,10 +54,10 @@
     "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^13.3.1",
-    "@angular-devkit/schematics": "^13.3.1",
-    "@angular/cli": "^13.3.1",
-    "@angular/compiler-cli": "^13.3.1",
+    "@angular-devkit/build-angular": "^14.2.13",
+    "@angular-devkit/schematics": "^14.2.13",
+    "@angular/cli": "^14.2.13",
+    "@angular/compiler-cli": "^14.3.0",
     "@types/jasmine": "^3.6.3",
     "@types/node": "^12.19.14",
     "eslint": "^8.13.0",
@@ -68,7 +68,7 @@
     "karma-coverage": "~2.0.3",
     "karma-jasmine": "~4.0.0",
     "karma-jasmine-html-reporter": "^1.5.0",
-    "ng-packagr": "^13.3.0",
+    "ng-packagr": "^14.2.2",
     "protractor": "~7.0.0",
     "ts-node": "~8.3.0",
     "typescript": "~4.6.3"
